+++ fltk-sys/build.rs
@@ -42,7 +42,7 @@ fn main() {
             .status()
             .expect("Curl and Tar are needed to download and upack the bundled libraries!");
     } else {
-        println!("cargo:rerun-if-changed=cfltk/include/cfl_new.hpp");
+        println!("cargo:rerun-if-changed=cfltk/CMakeLists.txt");
         println!("cargo:rerun-if-changed=cfltk/include/cfl.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_widget.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_group.h");
@@ -60,6 +60,8 @@ fn main() {
         println!("cargo:rerun-if-changed=cfltk/include/cfl_image.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_draw.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_table.h");
+        println!("cargo:rerun-if-changed=cfltk/include/cfl_surface.h");
+        println!("cargo:rerun-if-changed=cfltk/include/cfl_printer.h");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_new.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_widget.cpp");
@@ -79,7 +81,8 @@ fn main() {
         println!("cargo:rerun-if-changed=cfltk/src/cfl_draw.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_table.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_tree.cpp");
-        println!("cargo:rerun-if-changed=cfltk/CMakeLists.txt");
+        println!("cargo:rerun-if-changed=cfltk/src/cfl_surface.cpp");
+        println!("cargo:rerun-if-changed=cfltk/src/cfl_printer.cpp");
 
         Command::new("git")
             .args(&["submodule", "update", "--init"])
@@ -162,8 +165,15 @@ fn main() {
             handle_android(&target_triple, &mut dst);
         }
 
-        if target_triple.contains("linux" ) && !target_triple.contains("android") {
+        if target_triple.contains("linux") && !target_triple.contains("android") {
             dst.define("OPTION_USE_PANGO", "ON");
+        }
+
+        if target_triple.contains("unknown-linux-musl") {
+            dst.define("CMAKE_C_COMPILER", "musl-gcc");
+            dst.define("CMAKE_CXX_COMPILER", "musl-gcc");
+            dst.define("HAVE_STRLCPY", "False");
+            dst.define("HAVE_STRLCAT", "False");
         }
 
         let _dst = dst
@@ -176,7 +186,6 @@ fn main() {
             .define("OPTION_LARGE_FILE", "ON")
             .define("OPTION_BUILD_HTML_DOCUMENTATION", "OFF")
             .define("OPTION_BUILD_PDF_DOCUMENTATION", "OFF")
-            // .define("OPTION_ABI_VERSION:STRING", "10401")
             .build();
     }
 
@@ -278,6 +287,9 @@ fn main() {
                 println!("cargo:rustc-link-lib=dylib=fontconfig");
                 println!("cargo:rustc-link-lib=dylib=pango-1.0");
                 println!("cargo:rustc-link-lib=dylib=pangoxft-1.0");
+                println!("cargo:rustc-link-lib=dylib=gobject-2.0");
+                println!("cargo:rustc-link-lib=dylib=cairo");
+                println!("cargo:rustc-link-lib=dylib=pangocairo-1.0");
             }
         }
     }