+++ fltk-sys/cfltk/src/cfl_box.cpp
@@ -1,5 +1,4 @@
 #include "cfl_box.h"
-#include "cfl_new.hpp"
 #include <FL/Fl.H>
 #include <FL/Fl_Box.H>
 #include <FL/Fl_Image.H>