+++ fltk-sys/cfltk/src/cfl_input.cpp
@@ -1,5 +1,4 @@
 #include "cfl_input.h"
-#include "cfl_new.hpp"
 #include <FL/Fl.H>
 #include <FL/Fl_File_Input.H>
 #include <FL/Fl_Float_Input.H>