+++ fltk-sys/src/lib.rs
@@ -21,6 +21,8 @@ pub mod tree;
 pub mod valuator;
 pub mod widget;
 pub mod window;
+pub mod surface;
+pub mod printer;
 
 #[cfg(test)]
 mod tests {