+++ fltk/examples/image.rs
@@ -48,20 +48,19 @@
 
 use fltk::{app::*, frame::*, image::*, window::*};
 use std::error::Error;
-use std::path::*;
 
 fn main() -> Result<(), Box<dyn Error>> {
     let app = App::default().with_scheme(Scheme::Gleam);
     let mut wind = Window::new(100, 100, 400, 300, "Hello from rust");
     let mut frame = Frame::new(0, 0, 400, 300, "");
 
-    let mut image = SharedImage::load(&Path::new("screenshots/calc.jpg"))?;
+    let mut image = SharedImage::load("screenshots/calc.jpg")?;
     image.scale(200, 200, true, true);
 
     frame.set_image(Some(image));
 
-    // To remove an image
-    // frame.set_image(Option::<SharedImage>::None);
+    // // To remove an image
+    // frame.set_image(None::<SharedImage>);
 
     wind.make_resizable(true);
     wind.show();