+++ fltk/src/prelude.rs
@@ -29,6 +29,7 @@ pub enum FltkErrorKind {
     ResourceNotFound,
     ImageFormatError,
     TableError,
+    PrintError,
 }
 
 impl std::error::Error for FltkError {
@@ -964,4 +965,16 @@ pub unsafe trait ImageExt {
     unsafe fn delete(img: Self);
     /// Checks if the image was deleted
     fn was_deleted(&self) -> bool;
+}
+
+/// Defines the methods implemented by all surface types, currently ImageSurface
+pub trait SurfaceDevice {
+    /// Checks whether this surface is the current surface
+    fn is_current(&self) -> bool;
+    /// Get the current surface
+    fn surface() -> Self;
+    /// Push a surface as a current surface
+    fn push_current(new_current: &Self);
+    /// Pop the current surface
+    fn pop_current();
 }